remove packageinfo obj
